+++ lizard_layers/admin.py
@@ -25,6 +25,7 @@ class ParameterTypeAdmin(admin.ModelAdmin):
     list_display_links = ['value_type']
     list_display = [
         'value_type',
+        'measuring_rod',
         'parameter',
     ]